Rename folder CardDetails to PaymentForm
