+++ src/pages/Payment/PaymentDetails.tsx
@@ -1,7 +1,7 @@
 import Aos from "aos";
 import 'aos/dist/aos.css';
 
-import CardDetails from "../../components/PaymentPage/CardDetails";
+import PaymentForm from "../../components/Payment/PaymentForm";
 import { QuickInfo } from "../../components/SignUp";
 
 export const PaymentDetails = () => {
@@ -13,7 +13,7 @@ export const PaymentDetails = () => {
     return (
         <main>
             <div className="bg-gradient-to-b lg:space-x-[190px] md:px-20 px-10 lg:px-[130px] lg:flex lg:items-center from-[#F0F5FF] to-[#FFFFFF]">
-              <CardDetails />
+             <PaymentForm />
                 <QuickInfo />
             </div>
         </main>
